+++ packages/webview_flutter/webview_flutter_wkwebview/pigeons/web_kit.dart
@@ -747,6 +747,14 @@ abstract class WKUserContentController extends NSObject {
 abstract class WKPreferences extends NSObject {
   /// A Boolean value that indicates whether JavaScript is enabled.
   void setJavaScriptEnabled(bool enabled);
+
+  /// A Boolean value that indicates whether JavaScript can open windows
+  /// without user interaction.
+  ///
+  /// The default value is `false` on iOS and `true` on macOS.
+  ///
+  /// See https://developer.apple.com/documentation/webkit/wkpreferences/javascriptcanopenwindowsautomatically
+  void setJavaScriptCanOpenWindowsAutomatically(bool enabled);
 }
 
 /// An interface for receiving messages from JavaScript code running in a webpage.